- Update Dependencies & Migrate to Activity Compose 1.10.0-alpha03
This commit updates various dependencies to their latest versions, including:
- Room to 2.7.0-alpha11
- Navigation to 2.9.0-alpha02
- Lifecycle to 2.9.0-alpha06
- WorkManager to 2.10.0
- Activity Compose to 1.10.0-alpha03
- Compose BOM to 2024.10.01
- ConstraintLayout to 2.2.0
- Android Core to 1.15.0
- Fragment KTX to 1.8.5

It also migrates the codebase to use the new Activity Compose APIs, replacing `LocalActivity` with `androidx.activity.compose.LocalActivity` and adjusting usages accordingly. This ensures compatibility with the latest Compose updates and enables the use of newer features and improvements.