Refactor: Centralize Firebase ID creation
This commit refactors the way Firebase IDs are created and stored. Instead of setting them directly in each app's `onCreated` method, a new `createFirebaseIds` function is introduced in `OtakuApp`. This function returns a `FirebaseIds` data class containing all the necessary IDs, and these IDs are then set in `FirebaseDb` in the base `OtakuApp`.

- Adds a `FirebaseIds` data class to `OtakuApp` to hold Firebase ID values.
- Creates an abstract `createFirebaseIds()` function in `OtakuApp` to be implemented by child apps.
- Moves the logic for setting Firebase IDs to `OtakuApp.onCreate()` using the results of `createFirebaseIds()`.
- Updates `NovelApp`, `MangaApp`, and `AnimeApp` to implement `createFirebaseIds()` and remove the previous ID setting code from `onCreated()`.
- Marks `onCreate` as `@CallSuper` in `OtakuApp`.
- Creates an open `onCreated` function in `OtakuApp` to avoid breaking existing usage.