Temp: Add AndroidX snapshot repository for material-adaptive-navigation
This commit temporarily adds the AndroidX snapshot repository (`https://androidx.dev/snapshots/builds/13508953/artifacts/repository`) to `settings.gradle.kts`. This is necessary to use a snapshot version of `material-adaptive-navigation`.

A `TODO` comment and a logger error have been added to remind us to remove this repository once `material-adaptive-navigation` is no longer a snapshot.